fix: android old architecture
