+++ android/src/main/java/com/walletpayments/WalletPaymentsPackage.kt
@@ -1,13 +1,14 @@
 package com.walletpayments
 
-import com.facebook.react.BaseReactPackage
+import com.facebook.react.TurboReactPackage;
 import com.facebook.react.bridge.NativeModule
 import com.facebook.react.bridge.ReactApplicationContext
 import com.facebook.react.module.model.ReactModuleInfo
 import com.facebook.react.module.model.ReactModuleInfoProvider
+
 import java.util.HashMap
 
-class WalletPaymentsPackage : BaseReactPackage() {
+class WalletPaymentsPackage : TurboReactPackage() {
   override fun getModule(name: String, reactContext: ReactApplicationContext): NativeModule? {
     return if (name == WalletPaymentsModule.NAME) {
       WalletPaymentsModule(reactContext)
@@ -19,6 +20,7 @@ class WalletPaymentsPackage : BaseReactPackage() {
   override fun getReactModuleInfoProvider(): ReactModuleInfoProvider {
     return ReactModuleInfoProvider {
       val moduleInfos: MutableMap<String, ReactModuleInfo> = HashMap()
+      val isTurboModule = BuildConfig.IS_NEW_ARCHITECTURE_ENABLED
       moduleInfos[WalletPaymentsModule.NAME] = ReactModuleInfo(
         WalletPaymentsModule.NAME,
         WalletPaymentsModule.NAME,
@@ -26,7 +28,7 @@ class WalletPaymentsPackage : BaseReactPackage() {
         false,  // needsEagerInit
         true,  // hasConstants
         false,  // isCxxModule
-        true // isTurboModule
+        isTurboModule // isTurboModule
       )
       moduleInfos
     }
