--- android/src/newarch/WalletPaymentsSpec.kt
@@ -1,5 +1,0 @@
-package com.walletpayments
-
-import com.facebook.react.bridge.ReactApplicationContext
-
-abstract class WalletPaymentsSpec(context: ReactApplicationContext) : NativeWalletPaymentsSpec(context)
