+++ example/android/gradle.properties
@@ -37,5 +37,3 @@ newArchEnabled=true
 # Use this property to enable or disable the Hermes JS engine.
 # If set to false, you will be using JSC instead.
 hermesEnabled=true
-
-newArchEnabled=true
